Scan and Combine Files
This Program uses an ArrayList rather than a regular list in order to function. A regular Array has a fixed size that doesn't expand or decrease. However, an ArrayList can increase and decrease in length. This program scans files and adds each word to an ArrayList of ever-increasing size. The program does this for two objects of a Book class. As a result, there are now two arrayLists with separate books in each of them. The program then takes the odd number words from one book and combines them with the even number words from the second book. The final result is a FrankenNovel of two combined books. The book is called the Crimson Past, paying homage to The Scarlet Letter and The Great Gatsby.