Sand Art Project (Rotating Arrays)
In this project, I use linear algebra in order to flip, rotate, and invert Arrays. The purpose of this project is to be able to create a fully functional display of falling sand art pieces. The pieces are able to flip around the screen, mirror themselves, and even invert their own colors.